+++ src/components/profile/student/CourseOfStudySection.vue
@@ -12,12 +12,7 @@
             <div class="w-full lg:w-2/3">
                 <div class="mb-6 flex flex-col">
                     <label class="text-gray-700 text-md font-medium mb-3">Matriculation Number</label>
-                    <input
-                        type="text"
-                        readonly
-                        :value="matriculationId"
-                        class="w-full border-2 border-gray-400 rounded-lg py-3 text-gray-600 form-input bg-gray-300 focus:outline-none focus:shadow-none focus:border-gray-400"
-                    />
+                    <input type="text" readonly :value="matriculationId" class="w-full form-input input-text" />
                 </div>
                 <div class="mb-6 flex flex-col">
                     <label class="text-gray-700 text-md font-medium mb-3">Field of Study</label>
@@ -26,28 +21,18 @@
                             type="text"
                             readonly
                             :value="val"
-                            class="w-full border-2 border-gray-400 rounded-lg py-3 text-gray-600 form-input bg-gray-300 focus:outline-none focus:shadow-none focus:border-gray-400"
+                            class="w-full form-input input-text"
                             :class="{ 'mb-4': index !== fieldsOfStudy.length - 1 }"
                         />
                     </div>
                 </div>
                 <div class="mb-6 flex flex-col">
                     <label class="text-gray-700 text-md font-medium mb-3">Degree Sought</label>
-                    <input
-                        type="text"
-                        readonly
-                        value="Master's Degree"
-                        class="w-full border-2 border-gray-400 rounded-lg py-3 text-gray-600 form-input bg-gray-300 focus:outline-none focus:shadow-none focus:border-gray-400"
-                    />
+                    <input type="text" readonly value="Master's Degree" class="w-full form-input input-text" />
                 </div>
                 <div class="mb-6 flex flex-col">
                     <label class="text-gray-700 text-md font-medium mb-3">Semester Count</label>
-                    <input
-                        type="number"
-                        readonly
-                        :value="semesterCount"
-                        class="w-full border-2 border-gray-400 rounded-lg py-3 text-gray-600 form-input bg-gray-300 focus:outline-none focus:shadow-none focus:border-gray-400"
-                    />
+                    <input type="text" readonly :value="semesterCount" class="w-full form-input input-text" />
                 </div>
             </div>
         </div>